+++ item_cosmetic.rb
@@ -1,4 +1,4 @@
-class Cosmetic < Item
+class Cosmetic < ItemDecorator
   def initialize(args)
     @category = "cosmetic"
     super(args)
