+++ item_decorator.rb
@@ -1,0 +1,19 @@
+class ItemDecorator < Item
+  def initialize(args)
+    super(args)
+    @exempt    = exempt?
+    @imported  = imported? 
+    @taxation  = taxation
+  end
+
+  private
+
+    def exempt?
+      exempted_categories = ["food","medical","book"]
+      exempted_categories.include?(@category)     
+    end
+
+    def imported?
+      @name.include?("imported")
+    end
+end
