+++ item_media.rb
@@ -1,4 +1,4 @@
-class Media < Item
+class Media < ItemDecorator
   def initialize(args)
     @category = "media"
     super(args)
