+++ item_medical.rb
@@ -1,4 +1,4 @@
-class Medical < Item
+class Medical < ItemDecorator
   def initialize(args)
     @category = "medical"
     super(args)
